requirements.txt
django>=4.2
djangorestframework
psycopg2
django-cors-headers
requests
djangorestframework-simplejwt
cryptography
